+++ src/Symfony/Component/HttpClient/Tests/HttpClientBundleTest.php
@@ -24,6 +24,8 @@ use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigura
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBag;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\HttpClient\HttpClientBundle;
+use Symfony\Component\HttpClient\Recorder\RecorderConfigurationInterface;
+use Symfony\Component\HttpClient\RecorderHttpClient;
 use Symfony\Component\HttpClient\Response\MockResponse;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
@@ -113,6 +115,62 @@ class HttpClientBundleTest extends TestCase
         $this->assertFalse($container->hasDefinition('http_client'));
         $this->assertFalse($container->hasDefinition('http_client.transport'));
         $this->assertFalse($container->hasDefinition('cache.http_client'));
+    }
+
+    public function testTheRecorderIsWiredWhenEnabled()
+    {
+        $container = $this->loadExtension([
+            'default_options' => ['headers' => ['X-Foo' => 'bar']],
+            'recorder' => [
+                'enabled' => true,
+                'redact' => ['headers' => ['X-Custom-Secret'], 'query' => ['sig'], 'body' => ['pin']],
+            ],
+        ]);
+
+        $this->assertTrue($container->hasDefinition('http_client.recorder'));
+        $definition = $container->getDefinition('http_client.recorder');
+        $this->assertSame(RecorderHttpClient::class, $definition->getClass());
+
+        $this->assertSame(['http_client.transport', null, \PHP_INT_MAX - 1], $definition->getDecoratedService());
+
+        $arguments = $definition->getArguments();
+        $this->assertCount(6, $arguments);
+        $this->assertSame('.inner', (string) $arguments[0]);
+        $this->assertSame('http_client.recorder.store', (string) $arguments[1]);
+        $this->assertSame('http_client.recorder.configuration', (string) $arguments[2]);
+        $this->assertSame('http_client.recorder.matcher', (string) $arguments[3]);
+        $this->assertSame('http_client.recorder.redactor', (string) $arguments[4]);
+
+        // the recorder gets the same default options as the transport
+        $this->assertSame($container->getDefinition('http_client.transport')->getArgument(0), $arguments[5]);
+        $this->assertSame('bar', $arguments[5]['headers']['X-Foo'] ?? null);
+
+        $this->assertTrue($container->hasAlias(RecorderConfigurationInterface::class));
+        $this->assertSame([['X-Custom-Secret'], ['sig'], ['pin']], $container->getDefinition('http_client.recorder.redactor')->getArguments());
+        $this->assertSame('http_client.recorder.redactor', (string) $container->getDefinition('http_client.recorder.matcher')->getArgument(0));
+    }
+
+    public function testTheRecorderMatcherAndRedactorCanBeReplaced()
+    {
+        $container = $this->loadExtension([
+            'recorder' => ['enabled' => true, 'matcher' => 'my_matcher', 'redactor' => 'my_redactor'],
+        ]);
+
+        $this->assertSame('my_matcher', (string) $container->getAlias('http_client.recorder.matcher'));
+        $this->assertSame('my_redactor', (string) $container->getAlias('http_client.recorder.redactor'));
+    }
+
+    public function testTheRecorderIsDisabledByDefault()
+    {
+        $this->assertFalse($this->loadExtension([])->hasDefinition('http_client.recorder'));
+    }
+
+    private function loadExtension(array $config): ContainerBuilder
+    {
+        $container = new ContainerBuilder(new ParameterBag(['kernel.debug' => false]));
+        new HttpClientBundle()->getContainerExtension()->load([$config], $container);
+
+        return $container;
     }
 }